Implement config option to select tokenizer
